requirements.txt
torch
torchvision
torchaudio
numpy
pandas
matplotlib
scikit-learn
opencv-python
tqdm
requests
seaborn
pillow